+++ lighthouse-core/test/gather/gatherers/source-maps-test.js
@@ -12,75 +12,13 @@ jest.useFakeTimers();
 const Driver = require('../../../gather/driver.js');
 const Connection = require('../../../gather/connections/connection.js');
 const SourceMaps = require('../../../gather/gatherers/source-maps.js');
+const {createMockSendCommandFn, createMockOnFn,
+  flushAllTimersAndMicrotasks} = require('../mock-driver.js');
 const path = require('path');
 const fs = require('fs');
 const pathToMap = path.join(__dirname,
   '../../../../lighthouse-cli/test/fixtures/source-maps/bundle.js.map');
 const map = fs.readFileSync(pathToMap, 'utf-8');
-
-// Copied ...
-function createMockSendCommandFn() {
-  const mockResponses = [];
-  const mockFn = jest.fn().mockImplementation(command => {
-    const indexOfResponse = mockResponses.findIndex(entry => entry.command === command);
-    if (indexOfResponse === -1) throw new Error(`${command} unimplemented`);
-    const {response, delay} = mockResponses[indexOfResponse];
-    mockResponses.splice(indexOfResponse, 1);
-    if (delay) return new Promise(resolve => setTimeout(() => resolve(response), delay));
-    return Promise.resolve(response);
-  });
-
-  mockFn.mockResponse = (command, response, delay) => {
-    mockResponses.push({command, response, delay});
-    return mockFn;
-  };
-
-  mockFn.findInvocation = command => {
-    expect(mockFn).toHaveBeenCalledWith(command, expect.anything());
-    return mockFn.mock.calls.find(call => call[0] === command)[1];
-  };
-
-  return mockFn;
-}
-
-// new ...
-function createMockOnFn() {
-  const mockEvents = [];
-  const mockFn = jest.fn().mockImplementation((eventName, listener) => {
-    const events = mockEvents.filter(entry => entry.event === eventName);
-    if (!events.length) return;
-    for (const event of events) {
-      const indexOfEvent = mockEvents.indexOf(event);
-      mockEvents.splice(indexOfEvent, 1);
-    }
-    // Wait a tick because real events never fire immediately
-    setTimeout(() => {
-      for (const event of events) {
-        listener(event.response);
-      }
-    }, 0);
-  });
-
-  mockFn.mockEvent = (event, response) => {
-    mockEvents.push({event, response});
-    return mockFn;
-  };
-
-  mockFn.findListener = event => {
-    expect(mockFn).toHaveBeenCalledWith(event, expect.anything());
-    return mockFn.mock.calls.find(call => call[0] === event)[1];
-  };
-
-  return mockFn;
-}
-
-// Copied ...
-async function flushAllTimersAndMicrotasks() {
-  for (let i = 0; i < 1000; i++) {
-    jest.advanceTimersByTime(1);
-    await Promise.resolve();
-  }
-}
 
 describe('SourceMaps gatherer', () => {
   /**
